refactor: split message contract modules
