+++ src/messages.ts
@@ -1,486 +1,26 @@
 /**
- * popup ↔ background ↔ content script 間のメッセージ型定義。
- * すべてのメッセージはこの判別共用体のいずれかに該当する。
+ * Compatibility facade for runtime message contracts.
+ *
+ * Existing callers can continue importing from `src/messages`. New category
+ * modules own the definitions so each entrypoint can migrate independently.
  */
 
-import type { PlatformId } from './types/platform';
+import type { DiagnosticsMessage } from './messages/diagnostics';
+import type { MediaMessage } from './messages/media';
+import type { PostingMessage } from './messages/posting';
+import type { UpdateMessage } from './messages/update';
+import type { UserSessionMessage } from './messages/user-session';
 
 export type { PlatformId } from './types/platform';
-
-export type UserActionCategory =
-  | 'sign-in'
-  | 'check-account'
-  | 'complete-captcha'
-  | 'complete-confirmation'
-  | 'activate-tab'
-  | 'check-post-before-retry'
-  | 'fix-media'
-  | 'wait'
-  | 'report-ui-change';
-
-export type PostFlowStep = string;
-
-export interface PostFlowTrace {
-  mode?: 'preview' | 'post';
-  attempt?: string;
-  lastCompletedStep?: PostFlowStep;
-  failedStep?: PostFlowStep;
-  submitReached: boolean;
-  submissionStartedAt?: number;
-  tabUrlBefore?: string;
-  tabUrlAfter?: string;
-  urlCaptureTrace?: string[];
-}
-
-/**
- * 画像または動画の添付データ。
- * 拡張内 message では base64 文字列で運ぶ(ArrayBuffer は MV3 で潰れる)。
- * デコードは送信先側で行う(`src/utils/base64.ts` の `base64ToUint8Array`)。
- */
-export interface ImageAttachment {
-  name: string;
-  type: string; // MIME type
-  /**
-   * base64-encoded binary content。**dataRef がある場合は省略可** (sendMessage の
-   * 64MB 上限を超える binary は IndexedDB binary-transfer 経由で運ぶため)。
-   * popup の在庫状態 / 履歴では常に存在、message の wire format では dataRef
-   * 経由になり得る。
-   */
-  data?: string;
-  /**
-   * IndexedDB binary-transfer id。data 代わりに使う。
-   * `src/utils/binary-transfer.ts` の putBinary が返す id。受信側は
-   * `resolveAttachmentToBytes/Base64` で resolve してから使う。
-   */
-  dataRef?: string;
-  /** binary size (バイト)。base64 デコードしなくてもサイズが分かる */
-  bytes?: number;
-  /** 動画の場合の尺(秒)。background での制約チェックに使う */
-  durationS?: number;
-  /** 動画の場合の codec 名 (mediabunny: "avc", "hevc" など)。 */
-  videoCodec?: string;
-  /** MP4/WebM の codec parameter string (例: "avc1.64001f", "hev1.1.6.L93.90")。 */
-  videoCodecParameters?: string;
-  /**
-   * 画像 alt text (v0.4.87〜)。 Bluesky / Mastodon の API path で送信される。
-   * DOM 経路は対応してない (該当 SNS の compose UI に alt input が無い場合あり)。
-   */
-  alt?: string;
-}
-
-export type PostRequestIntent = 'new' | 'retry' | 'history-repost';
-export type SubmissionGuardDecision = 'allow' | 'blocked' | 'indeterminate';
-export type SubmissionGuardReason =
-  | 'in-flight'
-  | 'recent-success'
-  | 'recent-uncertain'
-  | 'fingerprint-unavailable'
-  | 'history-unavailable';
-
-export interface SubmissionGuardTrace {
-  decision: SubmissionGuardDecision;
-  reason?: SubmissionGuardReason;
-  requestId: string;
-}
-
-/** popup → background: 全 SNS への投稿リクエスト */
-export interface PostRequestMessage {
-  type: 'POST_REQUEST';
-  /** A unique identity for one user-triggered background submission attempt. */
-  requestId: string;
-  /** Why this request may create a post. Used by the background submission guard. */
-  intent: PostRequestIntent;
-  /** Required context for a History-originated repost. */
-  sourceHistoryEntryId?: string;
-  text: string;
-  platforms: PlatformId[];
-  images?: ImageAttachment[];
-  /**
-   * Request-time mode snapshot. Settings persistence is asynchronous in the
-   * popup, so background must not infer this only from storage.
-   * false = preview / dry-run, true = real post.
-   */
-  autoPost?: boolean;
-  /**
-   * v0.4.87〜: 詳細オプション。 SNS によって対応 / 無対応が分かれる:
-   * - cw (content warning / spoiler text): Mastodon (spoiler_text) / Misskey (cw)
-   * - visibility: Mastodon (public/unlisted/private/direct) / Misskey (public/home/followers/specified)
-   * - 他 SNS は ignore
-   */
-  cw?: string;
-  visibility?: 'public' | 'unlisted' | 'private' | 'direct';
-  /**
-   * 動画を先頭 N 秒で切り詰める (v0.4.90〜)。 user opt-in。
-   * popup の 「N 秒に切り詰めて投稿」 ボタンを user が明示クリックしたときに送信される。
-   */
-  trimVideoToSeconds?: number;
-}
-
-/** background → content script: 1 プラットフォームへの投稿指示 */
-export interface PostToPlatformMessage {
-  type: 'POST_TO_PLATFORM';
-  platform: PlatformId;
-  text: string;
-  /**
-   * thread mode: 文字数オーバーで分割されたチャンクを 1 つの compose 内に
-   * 連投スレッドとして並べる場合 (X のみ対応 v0.4.56〜)。指定時は text フィールド
-   * は無視され、textChunks の各要素が 1 ポストになる。images は最初のチャンクにだけ付く。
-   */
-  textChunks?: string[];
-  images?: ImageAttachment[];
-  /** dry-run: compose に流し込むが post button は押さない */
-  dryRun?: boolean;
-  /**
-   * popup / background が想定していた logged-in user (v0.4.83〜)。
-   * 例えば X で `@accountA` が前回検出されてれば `'@accountA'`。
-   * content script は post 直前に detectUser() で **現在 active な user** を
-   * 再検出し、 これと一致しなければ post を abort する (multi-account 誤爆防止)。
-   * undefined / null の場合は check skip (= detection 未実行 / 検出未対応 SNS)。
-   */
-  expectedUser?: string;
-  /** Mastodon / Misskey 用の CW / spoiler text (v0.4.87〜) */
-  cw?: string;
-  /** Mastodon / Misskey 用の visibility (v0.4.87〜) */
-  visibility?: 'public' | 'unlisted' | 'private' | 'direct';
-}
-
-/** content script → background: 1 プラットフォームの投稿結果 */
-export interface PostResultMessage {
-  type: 'POST_RESULT';
-  platform: PlatformId;
-  success: boolean;
-  /**
-   * Compose preview completed, but Tutti did not click the SNS post button.
-   * Preview results must never be stored as post history or carry post URLs.
-   */
-  preview?: boolean;
-  /**
-   * SNS 側で landing を確認できた場合のみ true。post URL、API の create 応答、
-   * または SNS 固有の完了 UI が証跡になる。preview=true の場合は compose の
-   * 準備完了を示すだけで、実投稿の landing 証跡ではない。
-   */
-  confirmed?: boolean;
-  /**
-   * Post click 後に証跡を取得できず、成功とも失敗とも断定できない状態。
-   * 非 idempotent な投稿を重複させないため、自動 retry 対象にしない。
-   */
-  uncertain?: boolean;
-  /**
-   * Background duplicate-policy result. Present when a platform is stopped
-   * before any posting side effect.
-   */
-  submissionGuard?: SubmissionGuardTrace;
-  /**
-   * Optional recovery category for popup guidance. This is intentionally coarse:
-   * the UI should tell the user what to do next without exposing brittle
-   * platform-internal step names.
-   */
-  userAction?: UserActionCategory;
-  flow?: PostFlowTrace;
-  error?: string;
-  /**
-   * 投稿後の post URL (= 「本当に landing した」 証跡)。 redirect 型 SNS では
-   * navigation 完了後の location、 API 型は public API から最新 post を引いた
-   * URL、 DOM scrape 型は自分の timeline から拾う。 取得経路がない / 失敗した
-   * 場合は省略 (success=true でも url なし状態あり、 popup は link なしで表示)。
-   */
-  url?: string;
-  /**
-   * post 後 verify 結果 (v0.4.75〜)。 post URL を取得した後、 SNS の public API
-   * もしくは DOM scrape で本文 / 画像 / tag が期待通りに書き込まれたかを確認した
-   * 結果。 verify が未対応 SNS では undefined、 verify error 時は verified=false。
-   */
-  verify?: {
-    verified: boolean;
-    issues: { kind: string; message: string; severity: 'warn' | 'error' }[];
-  };
-}
-
-/** background → popup: 1 プラットフォーム完了時の進捗通知(ストリーミング) */
-export interface PlatformProgressMessage {
-  type: 'PLATFORM_PROGRESS';
-  result: PostResultMessage;
-}
-
-export interface ExtensionUpdateState {
-  available: boolean;
-  version?: string;
-  applying?: boolean;
-}
-
-/** popup → background: downloaded extension update state を問い合わせる。 */
-export interface GetExtensionUpdateStateMessage {
-  type: 'GET_EXTENSION_UPDATE_STATE';
-}
-
-/** popup → background: user が選択したタイミングで downloaded update を適用する。 */
-export interface ApplyExtensionUpdateMessage {
-  type: 'APPLY_EXTENSION_UPDATE';
-}
-
-/** background → popup: Chrome が更新を download 済みにした。 */
-export interface ExtensionUpdateAvailableMessage {
-  type: 'EXTENSION_UPDATE_AVAILABLE';
-  state: ExtensionUpdateState;
-}
-
-/**
- * content script → background: 現在ログイン中のアカウント名。
- * v0.4.98: username に null を許容。 null は 「以前検出した値を clear する」 意味で、
- * 検出失敗 (detector が null 返却) の REFRESH_USER 応答で使う。
- */
-export interface CurrentUserMessage {
-  type: 'CURRENT_USER';
-  platform: PlatformId;
-  username: string | null;
-}
-
-// ── offscreen document 用メッセージ (P7: 動画整形) ──────────────────────────
-
-/** background → offscreen: 動画変換リクエスト */
-export interface ConvertVideoMessage {
-  type: 'CONVERT_VIDEO';
-  /**
-   * 入力動画の IndexedDB binary-transfer id (`src/utils/binary-transfer.ts`)。
-   * sendMessage の 64MB cap を回避するため拡張内 IndexedDB 経由で運ぶ。
-   * offscreen は getBinary(inputRef) で bytes を取得する。
-   */
-  inputRef: string;
-  /** mime type (e.g. "video/mp4")。出力は常に mp4/h.264/aac */
-  mimeType: string;
-  /** 動画長 (秒)。bitrate 計算に使う */
-  durationS: number;
-  /** 目標 byte 数。video bitrate = (targetBytes * 8 / duration) - audioKbps を解いて算出 */
-  targetBytes: number;
-  /**
-   * Aspect mode (v0.4.81〜):
-   * - `'passthrough'` (default): 元動画 aspect そのまま、 短辺 cap だけ適用 (854px)
-   * - `'vertical9x16'`: 1080×1920 (9:16) に letterbox + ぼかし背景。 TikTok / YT Shorts /
-   *   IG Reels 等の縦動画 SNS 向けの自動整形。 横長/正方形/縦長 すべてに対応
-   */
-  aspectMode?: 'passthrough' | 'vertical9x16';
-  /**
-   * 動画を先頭 N 秒で切り詰める (v0.4.90〜)。 user opt-in。
-   * CONCEPT 「尺超過は自動切り詰めせず拒否」 を破らないよう、 popup の「N 秒に切り詰めて投稿」
-   * ボタンを user が明示クリックしたときだけ送られる。 undefined / 0 なら切り詰めしない。
-   */
-  trimToSeconds?: number;
-}
-
-/** offscreen → background: 変換進捗 */
-export interface ConversionProgressMessage {
-  type: 'CONVERSION_PROGRESS';
-  progress: number; // 0-1
-  /** 進捗段階 (load / transcode) */
-  stage?: 'load' | 'transcode';
-}
-
-/** offscreen → background: 変換完了 (IndexedDB ref で返す) */
-export interface ConversionCompleteMessage {
-  type: 'CONVERSION_COMPLETE';
-  /** 出力動画の IndexedDB binary-transfer id。background が getBinary で取得 */
-  outputRef: string;
-  /** 出力後の byte 数 (popup 表示用) */
-  outputBytes: number;
-}
-
-/** offscreen → background: 変換エラー */
-export interface ConversionErrorMessage {
-  type: 'CONVERSION_ERROR';
-  error: string;
-}
-
-// ── 診断 (Diagnostics) ───────────────────────────────────────────────────────
-
-/** popup → background: 全 SNS の現在状態を吸い上げてレポート化する */
-export interface DiagnoseRequestMessage {
-  type: 'DIAGNOSE_REQUEST';
-  /** Error reports can limit diagnostics to the platforms involved in the failed draft. */
-  platforms?: PlatformId[];
-}
-
-/** background → content script: selector audit を要求 */
-export interface DiagnosePlatformMessage {
-  type: 'DIAGNOSE_PLATFORM';
-  platform: PlatformId;
-}
-
-/** content script → background: selector audit 結果 */
-export interface DiagnosePlatformResult {
-  type: 'DIAGNOSE_PLATFORM_RESULT';
-  platform: PlatformId;
-  url: string;
-  selectors: SelectorAudit[];
-  /** ログイン中ユーザー検出が成功したか(同じロジックで再実行) */
-  detectedUser: string | null;
-  /**
-   * 障害時の DOM 構造 snapshot (redacted)。selector miss が 1 件以上ある時のみ
-   * 含める。AI に「壊れたページの DOM はこれ、新 selector を提案して」と
-   * 渡す入力。本文 text / image src / value 属性は strip 済み。
-   */
-  domSnapshot: string | null;
-}
-
-export interface SelectorAudit {
-  /** 役割名(例: "fileInput", "textarea", "postButton") */
-  name: string;
-  selector: string;
-  matchCount: number;
-  /** 最初のマッチの outerHTML 先頭(短く) */
-  firstMatchPreview: string | null;
-}
-
-// ── Logger (各 context → background のログ集約) ───────────────────────────
-
-export type LogLevel = 'OFF' | 'ERROR' | 'WARN' | 'INFO' | 'DEBUG';
-
-export interface LogEntry {
-  /** 発行時刻 (ms epoch) */
-  ts: number;
-  level: LogLevel;
-  /** popup / background / SNS host 等の発行 context */
-  context: string;
-  message: string;
-}
-
-/** 任意 context → background: ログ 1 件追加 */
-export interface LogAppendMessage {
-  type: 'LOG_APPEND';
-  entry: LogEntry;
-}
-
-/** popup → background: バッファのスナップショット要求 */
-export interface LogExportRequestMessage {
-  type: 'LOG_EXPORT_REQUEST';
-}
-
-/** popup → background: バッファをクリア */
-export interface LogClearMessage {
-  type: 'LOG_CLEAR';
-}
-
-/**
- * popup → background: 圧縮 / 投稿の進行状態を問い合わせ。popup を閉じて
- * 再 open した時に、進行中の作業の UI を復活させるために使う。
- */
-export interface GetBgStateMessage {
-  type: 'GET_BG_STATE';
-}
-
-/**
- * popup → background: 完了済の posting state を消す (v0.4.96)。
- * postingStateInMemory を完了後も保持するようにした (popup 再 open で結果を
- * 復元する) ことに対するペアの explicit clear API。
- */
-export interface ClearPostingStateMessage {
-  type: 'CLEAR_POSTING_STATE';
-}
-
-/**
- * content script → background: 大きな binary を chunked で取得する。
- * tabs.sendMessage の 64MB cap を回避するため、media を base64 で全送りせず
- * dataRef を渡しておいて content script 側でこの message を loop 呼び出し。
- */
-export interface GetBinaryChunkMessage {
-  type: 'GET_BINARY_CHUNK';
-  dataRef: string;
-  offset: number;
-  length: number;
-}
-
-/**
- * background → bluesky content script: bsky.app の localStorage から
- * accessJwt / did / handle を抽出して返す。 reply chain で ATProto API path を
- * 使う際の認証 token を取得する用 (user が API credentials を設定してなくても、
- * 既に bsky.app にログインしてれば session を借りる)。
- */
-export interface GetBlueskySessionMessage {
-  type: 'GET_BLUESKY_SESSION';
-}
-
-/**
- * background → SNS content script: post detail page の og:* meta tag を
- * read back する (v0.4.77〜)。 og fetch (post-verify-og.ts) が login wall で
- * 失敗する SNS (X / IG / Threads etc) の DOM fallback。 logged-in tab で
- * meta tag を取れば SNS が auth 後 server-render する og:* を確実に拾える。
- */
-export interface VerifyPostDomMessage {
-  type: 'VERIFY_POST_DOM';
-}
-
-/**
- * popup / background → SNS content script: 当 tab で現在 active な user を
- * 即時 再検出して CURRENT_USER を送信させる (v0.4.83〜)。 popup が開かれたとき、
- * 各 SNS tab の現在 user を refresh して lastSeenUsers の stale を防ぐ
- * (multi-account 切替後 誤爆 guard と pair で使う)。
- */
-export interface RefreshUserMessage {
-  type: 'REFRESH_USER';
-}
-
-/**
- * popup → background: 全 SNS tab に REFRESH_USER を broadcast させる。
- * popup mount 時に呼ぶ。 background は adapter.matchUrl にマッチする tab を
- * 集めて REFRESH_USER を送る。
- */
-export interface BroadcastRefreshUsersMessage {
-  type: 'BROADCAST_REFRESH_USERS';
-}
-
-/** content script → background: captcha 等、利用者の画面操作が必要になった。 */
-export interface UserActionRequiredMessage {
-  type: 'USER_ACTION_REQUIRED';
-  platform: PlatformId;
-  reason: 'captcha' | 'confirmation';
-}
-
-export interface VerifyPostDomResult {
-  type: 'VERIFY_POST_DOM_RESULT';
-  ogDescription: string;
-  ogImage: string;
-  /** true when the hydrated post page exposes a non-avatar image/media element. */
-  hasImage?: boolean;
-  /** true when the hydrated post page exposes a video/player element. */
-  hasVideo?: boolean;
-  /** og:* が無い場合の fallback として、 ページ本文の text 抜粋 */
-  bodyExcerpt: string;
-}
-
-/** bsky.app localStorage から取り出した session 情報 (background 側で使う) */
-export interface BlueskySessionResult {
-  type: 'BLUESKY_SESSION_RESULT';
-  accessJwt: string;
-  did: string;
-  handle: string;
-  pdsHost?: string;
-}
+export * from './messages/diagnostics';
+export * from './messages/media';
+export * from './messages/posting';
+export * from './messages/update';
+export * from './messages/user-session';
 
 export type Message =
-  | PostRequestMessage
-  | PostToPlatformMessage
-  | PostResultMessage
-  | PlatformProgressMessage
-  | GetExtensionUpdateStateMessage
-  | ApplyExtensionUpdateMessage
-  | ExtensionUpdateAvailableMessage
-  | CurrentUserMessage
-  | ConvertVideoMessage
-  | ConversionProgressMessage
-  | ConversionCompleteMessage
-  | ConversionErrorMessage
-  | DiagnoseRequestMessage
-  | DiagnosePlatformMessage
-  | DiagnosePlatformResult
-  | LogAppendMessage
-  | LogExportRequestMessage
-  | GetBgStateMessage
-  | ClearPostingStateMessage
-  | GetBinaryChunkMessage
-  | GetBlueskySessionMessage
-  | BlueskySessionResult
-  | VerifyPostDomMessage
-  | VerifyPostDomResult
-  | RefreshUserMessage
-  | BroadcastRefreshUsersMessage
-  | UserActionRequiredMessage
-  | LogClearMessage;
+  | PostingMessage
+  | MediaMessage
+  | DiagnosticsMessage
+  | UpdateMessage
+  | UserSessionMessage;
